+++ src/routes/auth/profile.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { createFileRoute } from '@tanstack/react-router'
 
-export const Route = createFileRoute('/_auth/profile')({
+export const Route = createFileRoute('/auth/profile')({
   component: ProfileComponent,
 })
